fix: update tests for newer error handling

+++ packages/api/src/services/paymentMethodService.ts
@@ -6,28 +6,30 @@ import type {
   GetAllCustomerPaymentMethodsResponse,
   GetCustomerPaymentMethodResponse,
   OakClient,
+  Result,
 } from "../types";
 import { httpClient } from "../utils/httpClient";
 import { SDKError } from "../utils/errorHandler";
 import { buildQueryString } from "./helpers";
+import { err, ok } from "../types";
 
 export interface PaymentMethodService {
   add(
     customerId: string,
     paymentMethod: AddCustomerPaymentMethodRequest,
-  ): Promise<AddCustomerPaymentMethodResponse>;
+  ): Promise<Result<AddCustomerPaymentMethodResponse>>;
   get(
     customerId: string,
     paymentId: string,
-  ): Promise<GetCustomerPaymentMethodResponse>;
+  ): Promise<Result<GetCustomerPaymentMethodResponse>>;
   list(
     customerId: string,
     query?: GetAllCustomerPaymentMethodsQuery,
-  ): Promise<GetAllCustomerPaymentMethodsResponse>;
+  ): Promise<Result<GetAllCustomerPaymentMethodsResponse>>;
   delete(
     customerId: string,
     paymentMethodId: string,
-  ): Promise<DeletePaymentMethodResponse>;
+  ): Promise<Result<DeletePaymentMethodResponse>>;
 }
 
 export const createPaymentMethodService = (
@@ -36,26 +38,31 @@ export const createPaymentMethodService = (
   async add(
     customerId: string,
     paymentMethod: AddCustomerPaymentMethodRequest,
-  ): Promise<AddCustomerPaymentMethodResponse> {
+  ): Promise<Result<AddCustomerPaymentMethodResponse, SDKError>> {
     try {
       const token = await client.getAccessToken();
+      if (!token.ok) {
+        return err(token.error);
+      }
 
       const response = await httpClient.post<AddCustomerPaymentMethodResponse>(
         `${client.config.baseUrl}/api/v1/customers/${customerId}/payment_methods`,
         paymentMethod,
         {
           headers: {
-            Authorization: `Bearer ${token}`,
+            Authorization: `Bearer ${token.value}`,
           },
           retryOptions: client.retryOptions,
         },
       );
 
-      return response;
+      return ok(response);
     } catch (error) {
-      throw new SDKError(
-        `Failed to add payment method for customer ${customerId}`,
-        error,
+      return err(
+        new SDKError(
+          `Failed to add payment method for customer ${customerId}`,
+          error,
+        ),
       );
     }
   },
@@ -63,25 +70,30 @@ export const createPaymentMethodService = (
   async get(
     customerId: string,
     paymentId: string,
-  ): Promise<GetCustomerPaymentMethodResponse> {
+  ): Promise<Result<GetCustomerPaymentMethodResponse, SDKError>> {
     try {
       const token = await client.getAccessToken();
+      if (!token.ok) {
+        return err(token.error);
+      }
 
       const response = await httpClient.get<GetCustomerPaymentMethodResponse>(
         `${client.config.baseUrl}/api/v1/customers/${customerId}/payment_methods/${paymentId}`,
         {
           headers: {
-            Authorization: `Bearer ${token}`,
+            Authorization: `Bearer ${token.value}`,
           },
           retryOptions: client.retryOptions,
         },
       );
 
-      return response;
+      return ok(response);
     } catch (error) {
-      throw new SDKError(
-        `Failed to get payment method for customer ${customerId}`,
-        error,
+      return err(
+        new SDKError(
+          `Failed to get payment method for customer ${customerId}`,
+          error,
+        ),
       );
     }
   },
@@ -89,24 +101,29 @@ export const createPaymentMethodService = (
   async list(
     customerId: string,
     query?: GetAllCustomerPaymentMethodsQuery,
-  ): Promise<GetAllCustomerPaymentMethodsResponse> {
+  ): Promise<Result<GetAllCustomerPaymentMethodsResponse, SDKError>> {
     try {
       const token = await client.getAccessToken();
+      if (!token.ok) {
+        return err(token.error);
+      }
       const queryString = buildQueryString(query);
 
       const response =
         await httpClient.get<GetAllCustomerPaymentMethodsResponse>(
           `${client.config.baseUrl}/api/v1/customers/${customerId}/payment_methods${queryString}`,
           {
-            headers: { Authorization: `Bearer ${token}` },
+            headers: { Authorization: `Bearer ${token.value}` },
             retryOptions: client.retryOptions,
           },
         );
-      return response;
+      return ok(response);
     } catch (error) {
-      throw new SDKError(
-        `Failed to get payment method for customer ${customerId}`,
-        error,
+      return err(
+        new SDKError(
+          `Failed to get payment method for customer ${customerId}`,
+          error,
+        ),
       );
     }
   },
@@ -114,25 +131,30 @@ export const createPaymentMethodService = (
   async delete(
     customerId: string,
     paymentMethodId: string,
-  ): Promise<DeletePaymentMethodResponse> {
+  ): Promise<Result<DeletePaymentMethodResponse, SDKError>> {
     try {
       const token = await client.getAccessToken();
+      if (!token.ok) {
+        return err(token.error);
+      }
 
       const response = await httpClient.delete<DeletePaymentMethodResponse>(
         `${client.config.baseUrl}/api/v1/customers/${customerId}/payment_methods/${paymentMethodId}`,
         {
           headers: {
-            Authorization: `Bearer ${token}`,
+            Authorization: `Bearer ${token.value}`,
           },
           retryOptions: client.retryOptions,
         },
       );
 
-      return response;
+      return ok(response);
     } catch (error) {
-      throw new SDKError(
-        `Failed to delete payment method ${paymentMethodId} for customer ${customerId}`,
-        error,
+      return err(
+        new SDKError(
+          `Failed to delete payment method ${paymentMethodId} for customer ${customerId}`,
+          error,
+        ),
       );
     }
   },
